+++ CHANGELOG.md
@@ -3,4 +3,5 @@ v1.18.0
 - add atmospheric temperature //
 - cache atmospheric conductivity //
 - add heating/cooling aabbs
+- add melting and freezing //
 Not yet...